Moving the helper functions that create and write to arrays based on different value types from gm_graph to gm_util.
