+++ apps/output_cpp/gm_graph/src/gm_util.cc
@@ -1,4 +1,5 @@
 #include "gm_util.h"
+#include <assert.h>
 
 /*
  * This is an implementation of the Tokenizer class.
@@ -48,4 +49,35 @@ long Tokenizer::countNumberOfTokens() {
     }
     return number_of_tokens_;
 }
+
+/*
+ * Method to create arrays based on a given value type and size.
+ */
+void *getArrayType(VALUE_TYPE vt, int size) {
+    switch(vt) {
+        case GMTYPE_BOOL: return (void *) new bool[size];
+        case GMTYPE_INT: return (void *) new int[size];
+        case GMTYPE_LONG: return (void *) new long[size];
+        case GMTYPE_FLOAT: return (void *) new float[size];
+        case GMTYPE_DOUBLE: return (void *) new double[size];
+        case GMTYPE_END: assert(false); return NULL; // Control should never reach this case.
+    }
+    return NULL;
+}
+
+/*
+ * Method to write to a given location in an array based on the given value type.
+ * The value to be written is extracted from the given string 'val'.
+ */
+void storeValueBasedOnType(void *arr, long pos, std::string val, VALUE_TYPE vt) {
+    switch(vt) {
+        case GMTYPE_BOOL: ((bool *)arr)[pos] = (val == "true"); break;
+        case GMTYPE_INT: ((int *)arr)[pos] = atoi(val.c_str()); break;
+        case GMTYPE_LONG: ((long *)arr)[pos] = atol(val.c_str()); break;
+        case GMTYPE_FLOAT: ((float *)arr)[pos] = strtof(val.c_str(), NULL); break;
+        case GMTYPE_DOUBLE: ((double *)arr)[pos] = strtod(val.c_str(), NULL); break;
+        case GMTYPE_END: assert(false); return; // Control should never reach this case.
+    }
+}
+
