+++ apps/output_cpp/gm_graph/test/test_generic_adj_list.cc
@@ -53,17 +53,17 @@ int main(int argc, char** argv) {
 
     std::vector<VALUE_TYPE> vprop_schema;
     // format for sample.adj
-    vprop_schema.push_back(GMTYPE_FLOAT);
+//    vprop_schema.push_back(GMTYPE_FLOAT);
     // format for testgraph_generic.adj
-//    vprop_schema.push_back(GMTYPE_INT);
-//    vprop_schema.push_back(GMTYPE_BOOL);
+    vprop_schema.push_back(GMTYPE_INT);
+    vprop_schema.push_back(GMTYPE_BOOL);
     
     std::vector<VALUE_TYPE> eprop_schema;
     // format for sample.adj
-    eprop_schema.push_back(GMTYPE_DOUBLE);
+//    eprop_schema.push_back(GMTYPE_DOUBLE);
     // format for testgraph_generic.adj
-//    eprop_schema.push_back(GMTYPE_INT);
-//    eprop_schema.push_back(GMTYPE_FLOAT);
+    eprop_schema.push_back(GMTYPE_INT);
+    eprop_schema.push_back(GMTYPE_FLOAT);
 
     std::vector<void*> vertex_props;
     std::vector<void*> edge_props;
